chore: Refactor and center canvas
Refactored the project structure and centered the canvas, also removes
the rest of the interface like the logo.
