--- client/my-app/src/Canvas.css
@@ -1,5 +1,0 @@
-.drawing-board {
-    border-style: solid;
-    border-color: black;
-    border-width: 5px;
-}
