+++ client/src/components/App/App.css
@@ -1,0 +1,2 @@
+.App {
+}
