+++ client/src/components/App/App.js
@@ -1,0 +1,16 @@
+import React, { Component } from "react";
+import Canvas from "../Canvas/Canvas";
+
+import "./App.css";
+
+class App extends Component {
+render() {
+    return (
+      <div className="App">
+        <Canvas/>
+      </div>
+    );
+  }
+}
+
+export default App;
