+++ client/src/components/Canvas/Canvas.css
@@ -1,0 +1,15 @@
+.drawing-board {
+  padding: 0;
+  margin: auto;
+  display: block;
+  width: 800px;
+  height: 600px;
+  position: absolute;
+  top: 0;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  border-style: solid;
+  border-color: black;
+  border-width: 5px;
+}
